Fix edge case 1 task

+++ 0x14-bit_manipulation/1-print_binary.c
@@ -10,7 +10,7 @@ void print_binary(unsigned long int n)
 {
 	/*unsigned int mask = 32768;*/
 	/*unsigned long int mask = 4611686018427387904;*/
-	unsigned long long int mask = 9223372036854775808U;
+	unsigned long int mask = 9223372036854775808U;
 	/*unsigned long long int mask = 18446744073709551615U;*/
 	/*unsigned long int mask = 1 << 63;*/
 	int flag = 0, counter = 0;
